pyproject.toml
[tool.poetry]
name = "listings"
version = "0.1.0"
description = ""
authors = ["terenced"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
django-localflavor = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
